Small style improvements
git-svn-id: https://svn.sosy-lab.org/software/cpachecker/branches/predicate-memory-functions@43755 4712c6d2-40bb-43ae-aa4b-fec3f1bdfe4c

+++ src/org/sosy_lab/cpachecker/util/predicates/pathformula/pointeraliasing/AssignmentFormulaHandler.java
@@ -371,7 +371,7 @@ class AssignmentFormulaHandler {
     // to find out which ranges should be retained, we complement the assigned ranges
     // and union with the range encompassing completeRhsFormula [0, lhsBitSize)
     RangeSet<Long> retainedRangeSet =
-        lhsRangeSet.complement().subRangeSet(Range.closedOpen((long) 0, lhsBitSize));
+        lhsRangeSet.complement().subRangeSet(Range.closedOpen(0L, lhsBitSize));
 
     if (!retainedRangeSet.isEmpty()) {
       // there are some retained bits from previous LHS
